requirements.txt
python-dotenv
mysql-connector